+++ ImageFeed/API/Errors.swift
