+++ ImageFeed/API/ProfileService/ProfileService.swift
@@ -9,13 +9,6 @@ import Foundation
 
 final class ProfileService {
     
-    struct ProfileResult: Decodable {
-        let username: String
-        let firstName: String
-        var lastName: String?
-        var bio: String?
-    }
-    
     private init(){}
     private var task: URLSessionTask?
     private let urlSession = URLSession.shared
