+++ ImageFeed/Application/SceneDelegate.swift
