+++ ImageFeed/ImagesList/Models/ImagesListCellModel.swift
@@ -1,0 +1,14 @@
+//
+//  ImagesListCellModel.swift
+//  ImageFeed
+//
+//  Created by Вадим Шишков on 23.07.2023.
+//
+
+import Foundation
+
+struct ImagesListCellModel {
+    let imageURL: String
+    let imageIsLiked: Bool
+    let date: Date?
+}
