+++ ImageFeed/Resources/Info.plist
